Move rename 'pdb_tools.sifts' to 'structure_tools.sifts'.
