--- tests/conftest.py
@@ -1,2 +1,0 @@
-def pytest_addoption(parser):
-    parser.addoption("--quick", action="store_true", help="Run only quick tests.")
