+++ tests/test_structure_tools/test_sifts/residue_data_1.json
